+++ lib/Animation/Env.hs
@@ -4,30 +4,30 @@ import Animation.Type (UserInput(..))
 
 data Env =
     Env
-        { title       :: String
-        , fps         :: Int
-        , size        :: (Int, Int)
-        , velocity    :: Int
-        , baselength  :: Int
-        , bricklength :: Int
-        , numOfBricks :: Int
-        , posOfBricks :: [Int]
-        , lifes       :: Int
-        , wallsHeight :: Int
-        , wallsGap    :: Int
+        { title       :: String     -- ^ Title of the Game
+        , fps         :: Int        -- ^ Frames per second
+        , size        :: (Int, Int) -- ^ Size of the game
+        , velocity    :: Int        -- ^ Speed of the ball
+        , baselength  :: Int        -- ^ Length of the base (%)
+        , bricklength :: Int        -- ^ Length of the bricks
+        , numOfBricks :: Int        -- ^ Number of bricks
+        , lifes       :: Int        -- ^ Life of the bricks
+        , maxBalls    :: Int        -- ^ Maximum number of balls
+        , multFreq    :: Int        -- ^ Frequency of multipliers
+        , multSlow    :: Int        -- ^ How slow the multiplier advances
         }
 
 defaultEnv :: Env
 defaultEnv =
-    Env { title       = "BRICK BREAKER VIDEOGAME"
+    Env { title       = "BRICK BREAKER VIDEOGAME (MULTI-BALL)"
         , fps         = 20
-        , size        = (75, 22)
+        , size        = (100, 22)
         , velocity    = 1
-        , baselength  = 15 * fst (size defaultEnv) `div` 100 -- ^ (%)
-        , bricklength = 5  * fst (size defaultEnv) `div` 100 -- ^ (%)     
+        , baselength  = 5 * fst (size defaultEnv) `div` 100     
+        , bricklength = 2
         , numOfBricks = 0
-        , posOfBricks = []
         , lifes       = 2
-        , wallsHeight = 60 * (snd (size defaultEnv) - 2) `div` 100 -- ^ (%)
-        , wallsGap    = 20 *  fst (size defaultEnv)      `div` 100 -- ^ (%)
+        , maxBalls    = 150
+        , multFreq    = 3
+        , multSlow    = 5
         }