+++ lib/Animation/State.hs
@@ -5,6 +5,7 @@ import Control.Monad.Trans.Reader (ask)
 import Control.Monad.Trans.Class (lift)
 import System.IO (hReady, Handle(..), stdin, hSetEcho, hSetBuffering, BufferMode(NoBuffering))
 import Data.Char (toUpper, toLower)
+import Data.List (nubBy)
 
 import Animation.Env (Env(..))
 import Animation.Type ( Animation
@@ -17,6 +18,7 @@ data Direction
     = Positive
     | Negative
     | Neutral
+    deriving Eq
 
 directionFromInt :: Int -> Direction
 directionFromInt 0 = Neutral
@@ -31,21 +33,23 @@ directionToMultiplier Neutral  =  0
 
 data St =
     St
-        { position     :: (Int, Int)
-        , direction    :: (Direction, Direction)
+        { position     :: [(Int, Int)]
+        , direction    :: [(Direction, Direction)]
         , bXPosition   :: Int
         , bricks       :: [Object]
-        , walls        :: Maybe Object
+        , multiplier   :: [Object]
         , points       :: Int
         , userInputs   :: [UserInput]
         , status       :: GameStatus
+        , clock        :: Int
         }
 
--- | Allocation of the list of reduced positions in the game
--- | A reduced position is a 'x' value divided by the brick length
--- | Positions in this function are a list of 'x positions'. This means that
--- | given width = 50 then positions 49, 50, 51, 52,... correspond to points (49,0), (50,0), (1,1), (2,1),...
-
+{-|
+  Allocation of the list of reduced positions in the game
+  A reduced position is a 'x' value divided by the brick length
+  Positions in this function are a list of 'x positions'. This means that
+  given width = 50 then positions 49, 50, 51, 52,... correspond to points (49,0), (50,0), (1,1), (2,1),...
+-}
 bricksInPlace :: Int -> [Int] -> Int -> Int -> [Object]
 bricksInPlace width positions life bricklength = map (\x -> Brick (findPosition (bricklength*x) width 0) life) positions
            where findPosition x width level = if x < width then (x,level) else findPosition (x - width) width (level + 1)
@@ -53,7 +57,7 @@ bricksInPlace width positions life bricklength = map (\x -> Brick (findPosition 
 -- | Default state
 
 defaultSt :: St
-defaultSt = St (0, 0) (Neutral, Neutral) 0 [] Nothing 0 [] Stopped
+defaultSt = St [] [] 0 [] [] 0 [] Stopped 0
 
 -- | Management of the User Input
 
@@ -75,9 +79,7 @@ getUserInput = go Nothing
                                                 | x `isChar` 'q' = Stop      
                                                 | x `isChar` 's' = Start     
                                                 | x `isChar` 'r' = RestAuto   
-                                                | x `isChar` ' ' = RestAuto   
-                                                | x `isChar` 'j' = LeftWall  
-                                                | x `isChar` 'l' = RightWall 
+                                                | x `isChar` ' ' = RestAuto
                                                 | otherwise      = Undefined 
                             isChar c1 c2 = c1 == toLower c2 || c1 == toUpper c2 
 
@@ -89,12 +91,12 @@ next = do
     lift ( put ( nextInternal env input prevSt ) )
 
 nextInternal :: Env -> Maybe UserInput -> St -> St
-nextInternal (Env _ _ (width, height) velocity baselength bricklength _ _ _ wallHeight wallsGap) 
+nextInternal (Env _ _ (width, height) velocity baselength bricklength _ _ maxBalls multFreq multSlow) 
              userInput
-             prevSt@(St (prevX, prevY) (prevXDir, prevYDir) prevBXPos prevBricks prevWalls prevPoints readInputs prevStatus)
+             prevSt@(St prevPositions prevDirections prevBXPos prevBricks prevMultipliers prevPoints readInputs prevStatus clock)
              =
     
- -- | Management of next state according to GameStatus and UserInput. UserInput is included in State to avoid the repetition delay issue.
+    -- Management of next state according to GameStatus and UserInput. UserInput is included in State to avoid the repetition delay issue.
    
     case prevStatus of
         Paused        -> case userInput of 
@@ -108,9 +110,8 @@ nextInternal (Env _ _ (width, height) velocity baselength bricklength _ _ _ wall
         Starting      -> case userInput of 
                               Just RestAuto -> prevSt {status = Restarting }
                               Just Start    -> prevSt {status = Playing    }     
-                              _             -> prevSt { position   = (newBXPos + div baselength 2, prevY)
+                              _             -> prevSt { position   = [(newBXPos + div baselength 2, snd $ head prevPositions)]
                                                       , bXPosition = newBXPos 
-                                                      , walls      = newWalls
                                                       , userInputs = newInputs }
         LevelComplete -> case userInput of 
                               Just RestAuto -> prevSt {status = Restarting}
@@ -121,44 +122,71 @@ nextInternal (Env _ _ (width, height) velocity baselength bricklength _ _ _ wall
                                 Just Pause     -> prevSt {status = Paused }
                                 Just RestAuto  -> prevSt {status = Auto   }
                                 _  -> St 
-                                           { position   = (newX, newY)
-                                           , direction  = (newXDir, newYDir)
+                                           { position   = newPositions
+                                           , direction  = newDirections
                                            , bXPosition = newBXPos
                                            , bricks     = newBricks
-                                           , walls      = newWalls
+                                           , multiplier = newMultipliers
                                            , points     = newPoints
                                            , userInputs = newInputs
                                            , status     = newStatus
+                                           , clock      = newClock
                                            }
                          else prevSt {status = LevelComplete }
         Auto          -> if prevBricks /= [] then
                             case userInput of 
-                                Just Stop  -> prevSt {status = Stopped    }
-                                Just Pause -> prevSt {status = Paused     }
+                                Just Stop    -> prevSt {status = Stopped    }
+                                Just Pause   -> prevSt {status = Paused     }
                                 _  -> St 
-                                           { position   = (newX, newY)
-                                           , direction  = (newXDir, newYDir)
+                                           { position   = newPositions
+                                           , direction  = newDirections
                                            , bXPosition = newBXPos
                                            , bricks     = newBricks
-                                           , walls      = newWalls
+                                           , multiplier = newMultipliers
                                            , points     = newPoints
                                            , userInputs = newInputs
                                            , status     = newStatus
+                                           , clock      = newClock
                                            }
                          else prevSt {status = LevelComplete }
 
     where
- 
- -- | New_Unbounded tells us which would be the position of the ball if there is no bounding
-   
-    newXUnbounded          = prevX + directionToMultiplier prevXDir * velocity
-    newYUnbounded          = prevY + directionToMultiplier prevYDir * velocity
 
- -- | Position control of the base limited by the width - Repeating Input interrupts the action
+ -- Adding positions and directions for freshly created balls
+
+    newBallsPos = foldr (\pos -> (++) $ take 3 $ repeat pos) [] prevPositions
+    newBallsDir = foldr (\dir -> (++) $ filter (/= dir) [ (xDir,yDir) | xDir <- [Positive, Negative], yDir <- [Positive, Negative] ]) [] prevDirections
+
+ -- Removing positions and directions for lost balls
+
+    removeBalls = filter (\pair -> snd (fst pair) < height )  -- CHANGE FOR THE EASY FUNCTION
+
+ -- New_Unbounded tells us which would be the position of the ball if there is no bounding. Also we need to know just edge bounded positions.
+   
+    newXUnbounded          = map (\xPair -> fst xPair + directionToMultiplier (snd xPair) * velocity) xPosDirPair 
+    newYUnbounded          = map (\yPair -> fst yPair + directionToMultiplier (snd yPair) * velocity) yPosDirPair
+
+    newUnboundedPositions  = zip newXUnbounded newYUnbounded
+
+ -- Following info is necessary to check if new position is empty. Problematic case: ███  « 
+ --                                                                                     ● «
+ --                                                                                     ███ 
+
+    newXBounced pair       = ( posX pair + directionToMultiplier (dirX pair) * velocity * (-1), posY pair + directionToMultiplier (dirY pair) * velocity )
+    newYBounced pair       = ( posX pair + directionToMultiplier (dirX pair) * velocity, posY pair + directionToMultiplier (dirY pair) * velocity * (-1) )
+
+ -- Zipping positions with directions for easier managing of values
+
+    xPosDirPair            = zip (map fst prevPositions) (map fst prevDirections)
+    yPosDirPair            = zip (map snd prevPositions) (map snd prevDirections)
+
+ -- Position control of the base limited by the width - Repeating Input interrupts the action
     
-    newBXPos = case prevStatus of 
-                    Auto -> restricted (prevX - div baselength 2)
-                    _    -> baseDecisionTree userInput readInputs (moveBase (-2)) (moveBase ( 2)) prevBXPos 
+    newBXPos   = case prevStatus of 
+                      Auto -> if 3 > abs ( prevBXPos + div baselength 2 - fst ( head prevPositions ) ) then
+                              restricted ( (fst $ head prevPositions) - div baselength 2)
+                              else moveBase $ 3 * ( signum $ fst ( head prevPositions ) - prevBXPos )
+                      _    -> baseDecisionTree userInput readInputs (moveBase (-2)) (moveBase ( 2)) prevBXPos 
 
     moveBase i = let newBxPos = prevBXPos + i in restricted newBxPos
 
@@ -168,158 +196,164 @@ nextInternal (Env _ _ (width, height) velocity baselength bricklength _ _ _ wall
                                then 0
                                else position
 
- -- | Detection of collision with the base
-    
-    baseCollision          = newY >= height - 2 && newBXPos <= newX && newX <= newBXPos + baselength
+ -- Detection of collision with the base
 
-    baseCornerCollision    = newY >= height - 2 
-                          && not baseCollision 
-                          && ( newBXPos              <= newX + newX - prevX 
-                          &&   newBXPos + baselength >= newX + newX - prevX )
+    baseCollision pair  = posY pair >= height - 2 && newBXPos <= posX pair && posX pair <= newBXPos + baselength      
 
- -- | Auxiliary functions to consider the length of a brick, not just their position
- -- | completePositions returns a list of occupied positions given a list of Bricks
+    baseCornerCollision pair = posY pair >= height - 2 
+                            && not ( baseCollision pair )
+                            && ( newBXPos              <= posX pair + directionToMultiplier (dirX pair) 
+                            &&   newBXPos + baselength >= posX pair + directionToMultiplier (dirX pair) )
+
+ -- Auxiliary functions to consider the length of a brick, not just their position
+ -- completePositions returns a list of occupied positions given a list of Bricks
     
     addPositions (u,v) brl = zip [u .. (u + brl - 1)] $ take brl $ repeat v
     completePositions      = foldl (\x y -> x ++ addPositions (brickPosition y) bricklength) []
     
- -- | Identification of the coordinate that will be impacted according to ball direction for three 
- -- | different cases: Collision with top or botton (brickCollisionY), collision with one side (brickCollisionX)   
- -- | or collision with a corner (cornerCollision)
+ -- Identification of the coordinate that will be impacted according to ball direction for three 
+ -- different cases: Collision with top or botton (brickCollisionY), collision with one side (brickCollisionX)   
+ -- or collision with a corner (cornerCollision)
 
-    targetX                = ( newX + directionToMultiplier prevXDir, newY)
-    targetY                = ( newX, newY + directionToMultiplier prevYDir)
-    cornerTarget           = ( newX + directionToMultiplier prevXDir
-                             , newY + directionToMultiplier prevYDir )
+    singleTargetX pair     = ( posX pair + directionToMultiplier (dirX pair), posY pair)
+    singleTargetY pair     = ( posX pair, posY pair + directionToMultiplier (dirY pair))
+    singleTatgetC pair     = ( posX pair + directionToMultiplier (dirX pair) 
+                             , posY pair + directionToMultiplier (dirY pair) )
 
-    impossibleXCollision   = elem (bouncedTargetX, snd targetY) $ completePositions prevBricks
-    impossibleYCollision   = elem (fst targetX, bouncedTargetY) $ completePositions prevBricks
+    brickCollisionX pair   = elem (singleTargetX pair) (completePositions prevBricks)
+    brickCollisionY pair   = elem (singleTargetY pair) (completePositions prevBricks) 
+    cornerCollision pair   = elem (singleTatgetC pair) (completePositions prevBricks)
 
-    bouncedTargetX         = newX + directionToMultiplier prevXDir * velocity * (-1)
-    bouncedTargetY         = newY + directionToMultiplier prevYDir * velocity * (-1)
+    bounceX pair          = brickCollisionX pair && newXPosAvailable pair
+    bounceY pair          = brickCollisionY pair && newYPosAvailable pair
+    bounceC pair          = cornerCollision pair && not (bounceX pair) && not (bounceY pair)
+                         || ( brickCollisionX pair && not (newXPosAvailable pair) )
+                         || ( brickCollisionY pair && not (newYPosAvailable pair) )
 
-    brickCollisionX        = elem targetX      $ completePositions prevBricks
-    brickCollisionY        = elem targetY      $ completePositions prevBricks
-    cornerCollision        = elem cornerTarget $ completePositions prevBricks
+    newXPosAvailable pair = not $ elem (newXBounced pair) (completePositions prevBricks)
+    newYPosAvailable pair = not $ elem (newYBounced pair) (completePositions prevBricks)
 
-    xBrickBounce           = brickCollisionX && not impossibleXCollision
-    yBrickBounce           = brickCollisionY && not impossibleYCollision
-    bounceBack             = cornerCollision && not xBrickBounce && not yBrickBounce
-                          || ( brickCollisionX && impossibleXCollision )
-                          || ( brickCollisionY && impossibleYCollision )
-                          || ( (newXUnbounded <= 0 || newXUnbounded >= width) && impossibleXCollision )
-                          || ( newYUnbounded <= 0 && impossibleYCollision )
+ -- Identification of the block that will be hit.
 
-    wallCollisionX         = case prevWalls of
-                                 Just wall -> wallCollision wall targetX
-                                 Nothing -> False
-    wallCollisionY         = case prevWalls of
-                                 Just wall -> wallCollision wall targetY
-                                 Nothing -> False
-    wallCornerCollision    = case prevWalls of
-                                 Just wall -> wallCollision wall cornerTarget && not wallCollisionX && not wallCollisionY
-                                 Nothing -> False
+    targetBrick pair      = if bounceX pair && bounceY pair then map identify $ [singleTargetX pair] ++ [singleTargetY pair]
+                       else if bounceX pair then map identify [singleTargetX pair]
+                       else if bounceY pair then map identify [singleTargetY pair]
+                       else if bounceC pair then if cornerCollision pair                                then map identify [singleTatgetC pair]
+                                            else if brickCollisionX pair && not (newXPosAvailable pair) then map identify $ [newXBounced pair] ++ [singleTargetX pair]
+                                                                                                        else map identify $ [newYBounced pair] ++ [singleTargetY pair]
+                                                                                                        else []
+
+    targetBricks           = foldl (++) [] $ map targetBrick $ newUnboundedPositionsDirectionsPairs 
+
+ -- Filters the only brick that is hit by the ball given a target position and a list of Bricks.
     
-    wallCollision (Wall w) target = let condition x = x == fst target 
-                                                   && snd target >= div (height - 2 - wallHeight) 2 
-                                                   && snd target <= div (height - 2 + wallHeight) 2
-                                     in case w of (Left  xPos) -> condition xPos 
-                                                  (Right xPos) -> condition xPos
+    identify target         = let result = filter (\u -> snd target == snd (brickPosition u) 
+                                                      && fst target -  fst (brickPosition u) < bricklength 
+                                                      && fst target -  fst (brickPosition u) >= 0          ) prevBricks
+                               in if result == [] then Nothing else Just $ head result
 
- -- | Update positions and directions for next state
+ -- Update positions and directions for next state
     
-    newX =
-        case prevXDir of
-            Neutral  ->     newXUnbounded
-            Positive -> min newXUnbounded width
-            Negative -> max newXUnbounded 0
-    newY =
-        case prevYDir of
-            Neutral  ->     newYUnbounded
-            Positive -> min newYUnbounded height
-            Negative -> max newYUnbounded 0
-    newXDir =
-        case prevXDir of
-            Neutral  -> Neutral
-            Positive -> if newXUnbounded >= width || wallCollisionX || wallCornerCollision || xBrickBounce || bounceBack || baseCornerCollision
-                        then Negative
-                        else Positive
-            Negative -> if newXUnbounded <= 0     || wallCollisionX || wallCornerCollision || xBrickBounce || bounceBack || baseCornerCollision
-                        then Positive
-                        else Negative
-    newYDir =
-        case prevYDir of
-            Neutral  -> Neutral
-            Positive -> if brickCollisionY        || wallCollisionY || wallCornerCollision || yBrickBounce || bounceBack || baseCornerCollision || baseCollision
-                        then Negative
-                        else Positive
-            Negative -> if newYUnbounded <= 0     || wallCollisionY || wallCornerCollision || yBrickBounce || bounceBack 
-                        then Positive
-                        else Negative
+    newX    = map ( \pair -> case dirX pair of
+                                Neutral  ->     (posX pair)
+                                Positive -> min (posX pair) width
+                                Negative -> max (posX pair) 0
+                  ) newUnboundedPositionsDirectionsPairs
+
+    newY    = map ( \pair -> case dirY pair of
+                                Neutral  ->     (posY pair)
+                                Positive -> min (posY pair) height
+                                Negative -> max (posY pair) 0
+                  ) newUnboundedPositionsDirectionsPairs
+
+    newXDir = map ( \pair -> case dirX pair of
+                                Neutral  -> Neutral
+                                Positive ->
+                                    if posX pair >= width  || baseCornerCollision pair || bounceX pair || bounceC pair
+                                    then Negative
+                                    else Positive
+                                Negative ->
+                                    if posX pair <= 0      || baseCornerCollision pair || bounceX pair || bounceC pair
+                                    then Positive
+                                    else Negative
+                  ) newUnboundedPositionsDirectionsPairs
+
+    newYDir = map ( \pair -> case dirY pair  of
+                                Neutral  -> Neutral
+                                Positive -> 
+                                    if posY pair >= height || baseCollision pair || baseCornerCollision pair || bounceY pair || bounceC pair
+                                    then Negative
+                                    else Positive
+                                Negative -> 
+                                    if posY pair <= 0      || bounceY pair || bounceC pair
+                                    then Positive
+                                    else Negative
+                  ) newUnboundedPositionsDirectionsPairs
+
+-- Definition of new Positions and Directions of balls making sure that they are not overlapping 
+
+    (newPositions, newDirections) = let newPositionsDirectionsPairs = take maxBalls $ nubBy (==) $ zip
+                                                                      ( zip newX    newY    ++ (if gotMultiplier then newBallsPos else []) ) 
+                                                                      ( zip newXDir newYDir ++ (if gotMultiplier then newBallsDir else []) )
+                                     in unzip newPositionsDirectionsPairs
+
+-- Creating pairs of position - direction for managment of new states
+
+    newUnboundedPositionsDirectionsPairs = removeBalls $ zip newUnboundedPositions prevDirections
+
+    (posX, posY, dirX, dirY) = (fst . fst , snd . fst, fst . snd, snd . snd)
     
- -- | Update status in case the player is unable to bounce back the ball
+ -- Update status in case the player is unable to bounce back the ball
     
-    newStatus = if newY == height then Stopped else prevStatus
+    newStatus = if length newPositions == 0 then Stopped else prevStatus
  
- -- | Update the score in case of any brick collision 
+ -- Update the score in case of any brick collision 
     
-    newPoints = (+) prevPoints $ fromEnum $ xBrickBounce || yBrickBounce || bounceBack
-
- -- | Identification of the block that will be hit
-
-    targetBricks = let identify target = filter (\u -> snd target == snd (brickPosition u) 
-                                       && fst target -  fst (brickPosition u) < bricklength 
-                                       && fst target -  fst (brickPosition u) >= 0          ) prevBricks
-                    in if xBrickBounce && yBrickBounce then identify targetX ++ identify targetY
-                  else if xBrickBounce                 then identify targetX
-                  else if                 yBrickBounce then identify targetY
-                  else if bounceBack                   then if ( cornerCollision && not xBrickBounce && not yBrickBounce )
-                                                            then identify cornerTarget
-                                                            else if impossibleXCollision
-                                                                 then identify targetX ++ identify (bouncedTargetX, snd targetY)
-                                                                 else identify targetY ++ identify (fst targetX, bouncedTargetY)
-                                                                 else []
-
- -- | Update the bricks state according to collisions. Brick disappears if life = 0
+    newPoints = prevPoints + (length prevBricks - length newBricks)
+    
+ -- Update the bricks state according to collisions. Brick disappears if life == 0
     
     newBricks = foldl (flip changeBricks) prevBricks targetBricks
 
- -- | Update the life of the bricks
+ -- Update the life of the bricks
     
-    changeBricks x bricks = let brickTail  = filter ((/=) (brickPosition x) . brickPosition) bricks
-                                brickHurt  = Brick (brickPosition x) (life x - 1)
-                             in if life x > 0 then brickHurt : brickTail else brickTail
+    changeBricks target bricks = case target of 
+                                    Just x -> let brickTail  = filter ((/=) (brickPosition x) . brickPosition) bricks
+                                                  brickHurt  = Brick (brickPosition x) (life x - 1)
+                                               in if life x > 0 then brickHurt : brickTail else brickTail
+                                    _      -> bricks
 
- -- | Update the state of the walls. Repeating Input interrupts the action
+ -- Read & record the UserInputs - Repeating Input interrupts the action
 
-    newWalls = if elem RightWall readInputs then
-                  case userInput of
-                       Just RightWall -> Nothing
-                       Just LeftWall  -> Just $ Wall $ Left  wallsGap
-                       _              -> Just $ Wall $ Right (width - wallsGap)
-          else if elem LeftWall  readInputs then
-                  case userInput of
-                       Just LeftWall  -> Nothing
-                       Just RightWall -> Just $ Wall $ Right (width - wallsGap)
-                       _              -> Just $ Wall $ Left  wallsGap
-             else case userInput of
-                       Just LeftWall  -> Just $ Wall $ Left  wallsGap
-                       Just RightWall -> Just $ Wall $ Right (width - wallsGap)
-                       _              -> Nothing
-
- -- | Read & record the UserInputs - Repeating Input interrupts the action
-
-    newInputs = let wallInputs = case newWalls of  
-                     Just (Wall (Left _))  -> [LeftWall] 
-                     Just (Wall (Right _)) -> [RightWall] 
-                     _ -> []
-                 in baseDecisionTree userInput readInputs (MoveLeft : wallInputs) (MoveRight : wallInputs) wallInputs
+    newInputs = baseDecisionTree userInput readInputs [MoveLeft] [MoveRight] []
                            
- -- | Decision helper function
+ -- Decision helper function
 
     baseDecisionTree uInput rInput efect1 efect2 noEfect = case uInput of 
                                            Just MoveLeft  -> if not $ elem MoveRight rInput || elem MoveLeft  rInput then efect1 else noEfect
                                            Just MoveRight -> if not $ elem MoveLeft  rInput || elem MoveRight rInput then efect2 else noEfect
-                                           _              -> if elem MoveLeft rInput && prevBXPos > 0 then efect1
-                                                               else if elem MoveRight rInput && prevBXPos + baselength < width then efect2
-                                                                 else noEfect
+                                           _              -> if       elem MoveLeft  rInput && prevBXPos > 0                  then efect1
+                                                              else if elem MoveRight rInput && prevBXPos + baselength < width then efect2
+                                                                                                                              else noEfect
+
+-- Management of the Multipliers: Creation, desrtuction and velocity
+
+    newMultipliers  = let removeMultipliers                     = filter ( \multi@(Multiplier (x,y)) -> y < height + 1 && not (baseMultiplier multi) )
+                          createABrickBonus (targetBrick : tBs) = case targetBrick of
+                                                                      Just (Brick position _) -> [Multiplier position]
+                                                                      _                       -> createABrickBonus tBs
+                          createABrickBonus  _                  = []
+                       in removeMultipliers $ map advanceInY prevMultipliers 
+                       ++ if newPoints `mod` multFreq == 0 && (length prevBricks /= length newBricks) then createABrickBonus targetBricks else []
+
+    advanceInY (Multiplier (x,y)) = let fallControl = if mod clock multSlow < 1 then 1 else 0 in Multiplier (x , y + fallControl )    
+
+-- Catching Multipliers
+
+    baseMultiplier (Multiplier (multX, multY)) = multY == height - 1 && newBXPos <= multX && multX <= newBXPos + baselength
+
+    gotMultiplier = length prevMultipliers > length ( filter (not . baseMultiplier) (map advanceInY prevMultipliers) )
+
+-- Keeping track of frames
+ 
+    newClock = clock + 1